only_if

+++ docker/run.sh
@@ -6,16 +6,20 @@ for f in ${METAFILES}; do
   # download data
   echo "Now working on ${f}:"
 
-  cd ${METADIR}
-  echo -e "\t...pulling metadata..."
-  wget --quiet -O ${f}.csv ${SOURCEURL}/bundles/${f}.csv
+  if [ -f "${METADIR}/${f}.csv" ]; then
+    echo -e "\t...metadata already exists, skipping..."
+  else
+    cd ${METADIR}
+    echo -e "\t...pulling metadata..."
+    wget --quiet -O ${f}.csv ${SOURCEURL}/bundles/${f}.csv
 
-  cd ${DATADIR}
-  echo -e "\t...pulling data..."
-  wget --quiet -O ${f}.tar.bz2 ${SOURCEURL}/bundles/${f}-bundle.tar.bz2
+    cd ${DATADIR}
+    echo -e "\t...pulling data..."
+    wget --quiet -O ${f}.tar.bz2 ${SOURCEURL}/bundles/${f}-bundle.tar.bz2
 
-  echo -e "\t...extracting data..."
-  bunzip2 -f ${f}.tar.bz2 && tar xf ${f}.tar --strip-components=2 && rm ${f}.tar
+    echo -e "\t...extracting data..."
+    bunzip2 -f ${f}.tar.bz2 && tar xf ${f}.tar --strip-components=2 && rm ${f}.tar
+  fi
 done
 
 cd ${INSTALLDIR}
